added docstrings

+++ src/prognosis_api.py
@@ -17,16 +17,34 @@ rd1 = redis.Redis(host='redis-db', port=6379, db=1)
 
 
 def get_method() -> dict:
+    """
+    Outputs the all data retrieved from the Redis database.
+
+    Args:
+        none
+    Returns:
+        data (dict): dictionary with all data from breast cancer database
+    """
     global rd0
     try:
+        if len(rd0.keys()) == 0:
+            return f'Error. Breast cancer data not loaded in\n', 404
         data = []
         for item in rd0.keys():
             data.append(json.loads(rd0.get(item)))
         return data
     except Exception as err:
-        return f'Error. Data not loaded in\n', 404
+        return f'Error. Breast cancer data not loaded in\n', 404
 
 def m_cases() -> int:
+    """
+    Returns number of malignant cases in the breast cancer database.
+
+    Args:
+        none
+    Returns:
+        val_m (int): number of malignant cases
+    """
     global rd0
     try:
         val_m = 0
@@ -36,9 +54,17 @@ def m_cases() -> int:
                 val_m += 1
         return val_m
     except Exception as err:
-        return f'Error. Data not loaded in\n', 404
+        return f'Error. Breast cancer data not loaded in\n', 404
 
 def b_cases() -> int:
+    """
+    Returns number of benign cases in the breast cancer database.
+
+    Args:
+        none
+    Returns:
+        val_m (int): number of malignant cases
+    """
     global rd0
     try:
         val_b = 0
@@ -48,10 +74,19 @@ def b_cases() -> int:
                 val_b += 1
         return val_b
     except Exception as err:
-        return f'Error. Data not loaded in\n', 404
+        return f'Error. Breast cancer data not loaded in\n', 404
     
 @app.route('/data', methods = ['POST', 'GET', 'DELETE'])
-def breast_cancer_data() -> dict:
+def breast_cancer_data() -> str:
+    """
+    Handles all available methods of 'POST', 'GET', and 'DELETE' that can be requested by the user to load, return, or delete the breast cancer data from the database. 
+
+    Args:
+        none
+    Returns:
+        (str): status info whether 'POST' or 'DELETE' method was executed
+        data (dict): dictionary of the breast cancer data  
+    """
     global rd0
     if request.method == 'POST':
         try:
@@ -63,39 +98,67 @@ def breast_cancer_data() -> dict:
                 id_num = row.get('ID Number')
                 rd0.set(id_num, json.dumps(row))
         except Exception as err:
-            return f'Error. Data not loaded in\n', 404
+            return f'Error. Breast cancer data not loaded in\n', 404
         return f'Data loaded in\n'
-
     elif request.method == 'GET':
         return get_method()
-
     elif request.method == 'DELETE':
+        if len(rd0.keys()) == 0:
+            return f'Error. Breast cancer data not loaded in\n', 404
         rd0.flushdb()
         return f'Data deleted\n'
-                          
     else:
         return f'No available method selected. Methods available: POST, GET, DELETE\n', 404
 
 
 @app.route('/id', methods = ['GET'])
-def cancer_case_id() -> dict:
-    if rd0.keys() == 0:
-        return f'Error. Data not loaded in\n', 404
+def cancer_case_id() -> list:
+    """
+    Returns all ID numbers representing each case within the breast cancer data.
+
+    Args:
+        none
+    Returns:
+        rd0.keys() (list): list of all ID numbers of the data  
+
+    """
+    if len(rd0.keys()) == 0:
+        return f'Error. Breast cancer data not loaded in\n', 404
     return rd0.keys()
 
 
 @app.route('/id/<id_num>', methods = ['GET'])
 def id_data(id_num: int) -> dict:
-    if rd0.keys() == 0:
-        return f'Error. Data not loaded in\n', 404
+    """
+    Outputs data associated with <id_num> case within the breast cancer data. 
+
+    Args:
+        id_num (int): ID number value
+    Returns:
+        id_data (dict): data associated with <id_num>
+    """
+    if len(rd0.keys()) == 0:
+        return f'Error. Breast cancer data not loaded in\n', 404
     try:
-        return json.loads(rd0.get(id_num))
+        id_data = json.loads(rd0.get(id_num))
+        return id_data
     except Exception as err:
         return f'Error. ID {id_num} not found in database\n', 404
 
 
 @app.route('/outcome', methods = ['GET'])
 def get_cases() -> dict:
+    """
+    Returns ID numbers and total number of cases associated with diagnosis of malignant or benign.
+
+    Args:
+        none
+    Returns:
+        cases (dict): ID numbers and total number of malignant and benign cases
+    """
+    global r0
+    if len(rd0.keys()) == 0:
+        return f'Error. Breast cancer data not loaded in\n', 404
     try:
         m_list = []
         b_list = []
@@ -110,50 +173,60 @@ def get_cases() -> dict:
         cases = {"Malignant": {"Total cases": val_m, "IDs": m_list}, "Benign": {"Total cases": val_b, "IDs": b_list}}
         return cases
     except Exception as err:
-        return f'Error. Data not loaded in\n', 404
+        return f'Error. Breast cancer data not loaded in\n', 404
 
 
 @app.route('/image', methods = ['POST', 'GET', 'DELETE'])
-def image():
+def image() -> str:
+    """
+    Handles all available methods of 'POST', 'GET', and 'DELETE' that can be requested by the user to load, return, or delete the graph image created utilizing the breast cancer data from the database. 
+    
+    Args:
+        none
+    Returns:
+        (str): status info whether 'POST' or 'DELETE' method was executed
+        image (png): image of the graph created utilizing the breast cancer data  
+    """
     global rd0, rd1
     if request.method == 'POST':
         graph_data = {}
         if len(rd0.keys()) == 0:
-            return f'Data not loaded into database\n', 404
+            return f'Breast cancer data not loaded in\n', 404
         val_m = m_cases()
         val_b = b_cases()
         graph_data = {'Malignant': val_m, 'Benign': val_b}
         diagnosis = list(graph_data.keys())
         num_diagnosis = list(graph_data.values())
         plt.bar(diagnosis, num_diagnosis, color = 'maroon', width = .4)
-        plt.xlabel("Prognosis")
+        plt.xlabel("Diagnosis")
         plt.ylabel("Number of Cases")
-        plt.title("Breast Cancer Cases Prognosis")
+        plt.title("Breast Cancer Cases Diagnosis")
         plt.savefig('./cancer_prognosis.png')
         image_data = open('./cancer_prognosis.png', 'rb').read()
         rd1.set('image', image_data)
         return f'Graph successfully saved\n'
-
     elif request.method == 'GET':
         if len(rd0.keys()) == 0:
-            return f'Data not loaded into database\n', 404
+            return f'Error. Breast cancer data not loaded in\n', 404
+        elif len(rd1.keys()) == 0:
+            return f'Error. Image not loaded in\n', 404
         try:
             image_data = rd1.get('image')
             if image_data is None:
-                return f'Error. No image found\n', 404
+                return f'Error. Image not loaded in\n', 404
             with open('./cancer_prognosis.png', 'wb') as image_file:
                 image_file.write(image_data)
             return send_file('./cancer_prognosis.png', mimetype = 'image/png', as_attachment = True)
         except Exception as err:
             return f'Error. Unable to fetch image\n', 404
-
     elif request.method == 'DELETE':
         try:
+            if len(rd1.keys()) == 0:
+                return f'Error. Image not loaded in', 404
             rd1.flushdb()
             return f'Image deleted\n'
         except Exception as err:
-            return f'Unable to delete image\n', 404
-
+            return f'Error. Unable to delete image\n', 404
     else:
         return f'No available method selected. Methods available: POST, GET, DELETE\n', 404
